seo: add generateMetadata to all major pages and split client components
